requirements.txt
aiogram==2.7
aiohttp==3.6.2
aioredis==1.3.1
python-dotenv~=0.15.0
mysql~=0.0.2
mysql-connector-python~=8.0.23